Added chunk observers for block collision
- When a block in a chunk is changed for which we have at least one
  physics body, we remove any existing physics bodies for that changed
  block so they're either removed right away, or regenerated at the next
  physics step

- Tweaked forces for movement, jumping, as well as friction and
  bounciness

- Fixed improperly sized and placed player collision bounding box

+++ src/physics/BlockCollision.cpp
@@ -6,6 +6,7 @@
 #include "world/chunk/Chunk.h"
 #include "world/block/BlockRegistry.h"
 
+#include "io/Format.h"
 #include <Logging.h>
 
 #include <mutils/time/profiler.h>
@@ -58,13 +59,34 @@ void BlockCollision::startFrame() {
     const glm::vec3 bodyPos = vec(bodyTrans.getPosition());
 
     // discard all bodies that are too far away
-    const size_t erased = std::erase_if(this->bodies, [&, bodyPos](const auto &item) {
-        const auto& [blockPos, info] = item;
-        // if too far away, we need to delete the physics body to ensure it's not dangling around
-        if(distance2(bodyPos, glm::vec3(blockPos)) > kBlockMaxDistance) {
-            if(std::holds_alternative<BlockBody>(info)) {
-                const auto &body = std::get<BlockBody>(info);
-                this->engine->world->destroyRigidBody(body.body);
+    {
+        LOCK_GUARD(this->bodiesLock, BodiesLock);
+        const auto erased = std::erase_if(this->bodies, [&, bodyPos](const auto &item) {
+            const auto& [blockPos, info] = item;
+            // if too far away, we need to delete the physics body to ensure it's not dangling around
+            if(distance2(bodyPos, glm::vec3(blockPos)) > kBlockMaxDistance) {
+                this->removeBlockBody(blockPos, false);
+                this->decrementChunkRefCount(blockPos);
+
+                return true;
+            }
+
+            return false;
+        });
+
+        if(erased) {
+            // Logging::trace("Removed {} block bodies due to distance", erased);
+        }
+    }
+
+    // remove chunk observers that are no longer needed
+    const auto erased = std::erase_if(this->chunkObservers, [&](const auto &item) {
+        const auto& [chunkPos, token] = item;
+
+        if(!this->activeChunks.contains(chunkPos)) {
+            auto chunk = this->engine->scene->getChunk(chunkPos);
+            if(chunk) {
+                chunk->unregisterChangeCallback(token);
             }
 
             return true;
@@ -74,7 +96,7 @@ void BlockCollision::startFrame() {
     });
 
     if(erased) {
-        Logging::trace("Removed {} block bodies due to distance", erased);
+        // Logging::trace("Removed {} chunk observers", erased);
     }
 }
 
@@ -83,17 +105,62 @@ void BlockCollision::startFrame() {
  */
 void BlockCollision::removeAllBlocks() {
     PROFILE_SCOPE(RemoveAllBlocks);
+    LOCK_GUARD(this->bodiesLock, BodiesLock);
 
     // destroy the physics bodies
     for(const auto &[pos, info] : this->bodies) {
-        if(std::holds_alternative<BlockBody>(info)) {
-            const auto &body = std::get<BlockBody>(info);
-            this->engine->world->destroyRigidBody(body.body);
-        }
+        this->removeBlockBody(pos);
+        this->decrementChunkRefCount(pos);
     }
 
     // clear em out
     this->bodies.clear();
+}
+
+/**
+ * If it exists, removes and dellocates any resources associated with a block's body.
+ *
+ * @note Assumes the body map lock is already taken.
+ */
+void BlockCollision::removeBlockBody(const glm::ivec3 &blockPos, const bool remove) {
+    if(!this->bodies.contains(blockPos)) return;
+    const auto &info = this->bodies[blockPos];
+
+    if(std::holds_alternative<BlockBody>(info)) {
+        const auto &body = std::get<BlockBody>(info);
+        this->engine->world->destroyRigidBody(body.body);
+    }
+
+    if(remove) {
+        this->bodies.erase(blockPos);
+    }
+}
+
+/**
+ * Removes the chunk reference count for the chunk containing the given block.
+ *
+ * This will fail if the given block has a zero ref count.
+ */
+void BlockCollision::decrementChunkRefCount(const glm::ivec3 &blockPos) {
+    glm::ivec2 chunkPos;
+    world::Chunk::absoluteToRelative(blockPos, chunkPos);
+
+    auto it = this->activeChunks.find(chunkPos);
+    XASSERT(it != this->activeChunks.end(), "Invalid ref count for chunk {} (block {})", chunkPos, blockPos);
+
+    this->activeChunks.erase(it);
+}
+
+/**
+ * Chunk block change callback
+ */
+void BlockCollision::chunkBlockDidChange(world::Chunk *chunk, const glm::ivec3 &blockCoord, const world::Chunk::ChangeHints hints) {
+    PROFILE_SCOPE(PhysicsChunkChangeCb);
+
+    const auto pos = blockCoord + glm::ivec3(256 * chunk->worldPos.x, 0, 256 * chunk->worldPos.y);
+
+    LOCK_GUARD(this->bodiesLock, BodiesLock);
+    this->removeBlockBody(pos);
 }
 
 /**
@@ -110,6 +177,8 @@ void BlockCollision::update() {
     const glm::ivec3 bodyPos = floor(vec(bodyTrans.getPosition()));
 
     // check the blocks around us
+    LOCK_GUARD(this->bodiesLock, BodiesLock);
+
     for(int y = -kLoadYRange; y <= kLoadYRange; y++) {
         for(int z = -kLoadXZRange; z <= kLoadXZRange; z++) {
             for(int x = -kLoadXZRange; x <= kLoadXZRange; x++) {
@@ -124,10 +193,20 @@ void BlockCollision::update() {
 
                 const auto chunk = scene->getChunk(chunkPos);
                 if(!chunk) {
-                    this->bodies[blockPos] = BlockNoCollision();
                     continue;
                 }
 
+                // set up handler for this block, if required
+                if(!this->activeChunks.contains(chunkPos)) {
+                    auto tok = chunk->registerChangeCallback(
+                            std::bind(&BlockCollision::chunkBlockDidChange, this,
+                                std::placeholders::_1, std::placeholders::_2, std::placeholders::_3));
+                    this->chunkObservers[chunkPos] = tok;
+                }
+
+                this->activeChunks.insert(chunkPos);
+
+                // get the block at this position
                 const auto block = chunk->getBlock(blockOff);
                 if(!block) {
                     this->bodies[blockPos] = BlockNoCollision();
@@ -144,6 +223,10 @@ void BlockCollision::update() {
 
                     Transform blockTransform(vec(this->blockTranslate), Quaternion::identity());
                     b.collider = b.body->addCollider(this->blockShape, blockTransform);
+
+                    auto &mat = b.collider->getMaterial();
+                    mat.setFrictionCoefficient(kFrictionCoefficient);
+                    mat.setBounciness(0);
 
                     this->bodies[blockPos] = b;
                 }
